Modularise backend
